Fixed previous commit

+++ contrib/gdevsvg.c
@@ -2240,7 +2240,7 @@ gx_image_enum_common_t ** pinfo)
 	}
 	code = 0;			/* for normal path */
 
-	int ppiDecode = 0.0; // Not sure if this is an appropriate default, have't needed to use the default yet in testing
+	float ppiDecode = 0.0; // Not sure if this is an appropriate default, have't needed to use the default yet in testing
 	if (ppi->Decode != NULL)
 	{
 		ppiDecode = ppi->Decode[0];
